[Kernels][GPU] Remove the int(tid) pattern, NFC (#51830)
We should not have to cast to an int to perform these operations.

NN_ORIG_REV_ID: b2f3435c957e293b473c87edc2ce5ccbcd15c379

+++ mojo/kernels/src/nn/toppminp_gpu.mojo
@@ -342,12 +342,12 @@ fn radix_sort_pairs_kernel[
             var key = input_keys[index]
             var normalized_key = normalize(key)
             var radix = (normalized_key >> current_bit) & (NUM_BUCKETS - 1)
-            counts[int(radix)] += 1
+            counts[radix] += 1
 
     # Store counts[NUM_BUCKETS] per thread into shared memory s_counts
     @parameter
     for i in range(NUM_BUCKETS):
-        s_counts[int(tid * NUM_BUCKETS + i)] = counts[i]
+        s_counts[tid * NUM_BUCKETS + i] = counts[i]
     barrier()
 
     # Compute total_counts[NUM_BUCKETS] by summing counts[NUM_BUCKETS] across threads
